WIP more authorization interface cleanup.
Still have problem in `methods.spec.ts`
